src/requirements.txt
matplotlib==3.10.8
nibabel==5.4.2
numpy==2.4.3
pandas==3.0.1
PyYAML==6.0.2
PyYAML==6.0.3
tqdm==4.66.6
